requirements.txt
# requirements.txt - 组合风控系统依赖包
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4
tushare==1.2.89
arch==6.2.0
streamlit==1.28.2
plotly==5.15.0
configparser==5.3.0
python-dotenv==1.0.0
matplotlib==3.7.2
